Update patch set 1
Patch Set 1:

I can confirm that Python3.9 is installed in this node. Here is the list of packages that ran for this particular node:
https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/dcaegen2-collectors-snmptrap-master-snmptrap-verify-python/4/packages_end.txt.gz

I wonder if the issue is because the virtual env is trying to run using Python2:
2:21:30 ERROR:   py39: InvocationError for command /w/workspace/dcaegen2-collectors-snmptrap-master-snmptrap-verify-python/venv-tox/bin/python2 -m virtualenv --no-download --python /usr/bin/python3.9 py39 (exited with code 1)

Is there a way that you guys can force venv-tox/bin/python3 or something like that?

Patch-set: 1
Reviewer: Gerrit User 481 <481@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>